+++ packages/fighting-design/textarea/src/textarea.vue
@@ -1,11 +1,9 @@
 <script lang="ts" setup>
   import { Props } from './props'
-  import { ref, watch, nextTick, onMounted, onBeforeUnmount } from 'vue'
+  import { ref, nextTick, onMounted } from 'vue'
   import { useInput, useList, useRun } from '../../_hooks'
   import { FIconCross } from '../../_svg'
   import { FSvgIcon } from '../../svg-icon'
-  import { debounce } from '../../_utils'
-  import type { WatchStopHandle } from 'vue'
 
   defineOptions({ name: 'FTextarea' })
 
@@ -30,11 +28,13 @@
 
   /** 改变高度 */
   const changeHeight = async (): Promise<void> => {
-    await nextTick()
-    if (textareaRef.value) {
-      textareaRef.value.style.height = 'auto'
-      textareaRef.value.style.height = textareaRef.value.scrollHeight + 'px'
+    if (!textareaRef.value || !prop.autoHeight) {
+      return
     }
+
+    await nextTick()
+
+    textareaRef.value.style.height = textareaRef.value.scrollHeight + 'px'
   }
 
   /**
@@ -42,57 +42,11 @@
    *
    * @param { Object } evt 事件对象
    */
-  const _handleInput = (evt: Event): void => {
+  const listerInput = (evt: Event): void => {
     handleInput(evt)
 
-    // 如果需要自适应高度，则每次输入的时候重新设置高度
-    if (prop.autoHeight) {
-      changeHeight()
-    }
+    changeHeight() // 如果需要自适应高度，则每次输入的时候重新设置高度
   }
-
-  /** 是否为自动高度 */
-  const isAutoHeight = (): void => {
-    /**
-     * 监视器实例
-     *
-     * 监视绑定值 行数 是否自动高度的数据变化，来设置文本框的高度
-     */
-    const unwatch: WatchStopHandle = watch(
-      (): (string | number | boolean)[] => [prop.modelValue, prop.rows, prop.autoHeight],
-      (): void => {
-        changeHeight()
-
-        // 如果中途自动高度为假了，则取消监视器
-        if (!prop.autoHeight) {
-          unwatch()
-          // 并且设置自动高度
-          if (textareaRef.value) {
-            textareaRef.value.style.height = 'auto'
-          }
-        }
-      },
-      {
-        immediate: true
-      }
-    )
-  }
-
-  /** 设置宽度方法 */
-  const _changeHeight = debounce(changeHeight, 500)
-
-  // 初始化调用
-  onMounted(() => {
-    if (prop.autoHeight) {
-      isAutoHeight()
-      window.addEventListener('resize', _changeHeight)
-    }
-  })
-
-  // 销毁前移除事件
-  onBeforeUnmount(() => {
-    window.removeEventListener('resize', _changeHeight)
-  })
 
   /**
    * 监听回车事件
@@ -106,19 +60,6 @@
   const handleEnterKey = (evt: KeyboardEvent): void => {
     // 如果按下 Enter 和 Ctrl 触发换行
     if (evt.key === 'Enter' && evt.ctrlKey) {
-      modelValue.value += '\n'
-
-      // 如果是自适应高度，则重新设置高度
-      if (prop.autoHeight) {
-        changeHeight()
-      }
-
-      // 返回阻止继续执行
-      return
-    }
-
-    // 只有在按下 Enter 才触发回调方法
-    if (evt.key === 'Enter') {
       /**
        * 阻止默认换行的事件
        *
@@ -129,6 +70,9 @@
       run(prop.onEnter, modelValue.value, evt)
     }
   }
+
+  // 初始化调用
+  onMounted(changeHeight)
 </script>
 
 <template>
@@ -143,7 +87,7 @@
       :autofocus
       :placeholder
       :name
-      @input="_handleInput"
+      @input="listerInput"
       @change="handleChange"
       @blur="onBlur"
       @focus="onFocus"