+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - 修复 `f-checkbox-group` 组件 `row-gap` 配置项失效问题
 - 禁止了 `f-form` 组件的原生表单校验
 - 修复 `f-textarea` 的 outline 样式不一致性
+- 修复 `f-textarea` 组件在 `resize` 不同配置项的情况下输入框错位的问题
+- 优化 `f-textarea` 组件整体流程度，代码近一步优化
+- `f-textarea` 组件 Enter 事件默认改为换行，ctrl + Enter 为提交事件，会执行 `on-enter` 事件
 
 ## 1.0.0-alpha.8 (2024-06-12)